refactor: :art: integrate embla carousel for sectors display
