+++ eslint.config.js
@@ -65,13 +65,13 @@ export default tseslint.config(
             ['external', 'builtin', 'tanstack', 'tabler-icons', 'lucide-icons'],
             'custom-features',
             'custom-routes',
+            'custom-types',
             'custom-hooks',
             'custom-utils',
             'internal',
             'custom-components',
             'custom-pages',
             'custom-auth',
-            'custom-types',
             'style',
             ['parent', 'sibling', 'index'],
             ['parent-type', 'sibling-type', 'index-type'],
