+++ package.json
@@ -34,7 +34,9 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "1.0.0",
+    "embla-carousel-autoplay": "^8.5.1",
     "embla-carousel-react": "^8.5.1",
+    "embla-carousel-wheel-gestures": "^8.0.1",
     "eslint-plugin-react": "^7.37.2",
     "framer-motion": "^11.12.0",
     "i18next": "^24.0.2",
